IT WORKS!
All right, it's actually working and seems to be robust!  Added in
another c81 file and did some basic tests.  Everything seems to be good.
Partially incorporated the BEMT into vehicle.py, but it doesn't use all
the config values at the moment, so that'll have to be fixed, and
everything will have to be switched over to BEMT and momentum theory
removed.  It even runs quickly!  I'd like to tweak the tolerances and
differentials (or better yet, put them in the config) to try to improve
speed even further.  But the main thing is fully integrating it and
making sure all the config values are consistent etc., plus doing full
validation.